update through files for moves etc
