+++ neurodsp/spectral/__init__.py
@@ -5,4 +5,4 @@ from .power import (compute_spectrum, compute_spectrum_welch, compute_spectrum_w
 from .measures import compute_absolute_power, compute_relative_power, compute_band_ratio
 from .variance import compute_scv, compute_scv_rs, compute_spectral_hist
 from .utils import trim_spectrum, trim_spectrogram
-from ..sim.utils import rotate_spectrum as rotate_powerlaw
+from ..sim.modulate import rotate_spectrum as rotate_powerlaw
